refactor(skills): split organization members table
